rename: 페이지 파일 ui 폴더 안으로 이동
